[Unwanted] Remove Local Node Integration
This integration assumes the node is in the same machine and
tries to connect with it through rpc and p2p while modifies
the original bitcoin.conf file.

It is mostly used by integration testing and while only very few
people use it, it only provides blocks and fee estimations.

This commit moves the code useful for testing to the testing
project and removes the rest.